+++ src/ray/gcs/BUILD.bazel
@@ -33,6 +33,7 @@ ray_cc_library(
         ":gcs_table_storage",
         "//src/ray/common:asio",
         "//src/ray/common:id",
+        "//src/ray/gcs/postable",
         "//src/ray/protobuf:gcs_cc_proto",
         "@com_google_absl//absl/container:flat_hash_map",
     ],
@@ -46,6 +47,7 @@ ray_cc_library(
         "//src/ray/common:asio",
         "//src/ray/common:status",
         "//src/ray/gcs:grpc_service_interfaces",
+        "//src/ray/gcs/postable",
         "//src/ray/protobuf:gcs_cc_proto",
     ],
 )
@@ -130,6 +132,7 @@ ray_cc_library(
     hdrs = ["store_client_kv.h"],
     deps = [
         ":gcs_kv_manager",
+        "//src/ray/gcs/postable",
         "//src/ray/gcs/store_client",
     ],
 )
@@ -141,6 +144,7 @@ ray_cc_library(
     deps = [
         "//src/ray/common:asio",
         "//src/ray/gcs:grpc_service_interfaces",
+        "//src/ray/gcs/postable",
         "//src/ray/protobuf:gcs_service_cc_proto",
         "//src/ray/pubsub:gcs_publisher",
         "@com_google_absl//absl/container:flat_hash_map",
