+++ src/ray/gcs/gcs_kv_manager.h
@@ -20,9 +20,9 @@
 #include <vector>
 
 #include "ray/common/asio/instrumented_io_context.h"
-#include "ray/common/asio/postable.h"
 #include "ray/common/status.h"
 #include "ray/gcs/grpc_service_interfaces.h"
+#include "ray/gcs/postable/postable.h"
 
 namespace ray {
 namespace gcs {
