+++ src/ray/gcs/store_client/redis_store_client.h
@@ -24,7 +24,7 @@
 
 #include "absl/synchronization/mutex.h"
 #include "ray/common/asio/instrumented_io_context.h"
-#include "ray/common/asio/postable.h"
+#include "ray/gcs/postable/postable.h"
 #include "ray/gcs/store_client/redis_context.h"
 #include "ray/gcs/store_client/store_client.h"
