+++ src/ray/util/tests/BUILD.bazel
@@ -11,16 +11,6 @@ ray_cc_test(
 )
 
 ray_cc_test(
-    name = "function_traits_test",
-    srcs = ["function_traits_test.cc"],
-    tags = ["team:core"],
-    deps = [
-        "//src/ray/util:function_traits",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-ray_cc_test(
     name = "thread_checker_test",
     size = "small",
     srcs = ["thread_checker_test.cc"],
